feat: refactor Header component and remove HeaderUser
